Revert "sobec/contact/fwd.hpp"
This reverts commit 8e3d492a877bb75d4fc2c8c0a16b7607559391e3.
