+++ include/sobec/contact/contact1d.hpp
@@ -18,7 +18,6 @@
 #include <pinocchio/multibody/data.hpp>
 #include <pinocchio/spatial/motion.hpp>
 
-#include "sobec/contact/fwd.hpp"
 #include "sobec/fwd.hpp"
 
 namespace sobec {
